LayoutTests/storage/indexeddb/mozilla/delete-result-expected.txt

Test IndexedDB: result property after deleting existing and non-existing record

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true

indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo', { keyPath: 'id', autoIncrement: true });
request = objectStore.add({});
id = event.target.result;
request = objectStore.delete(id);
PASS event.target.result is undefined
request = objectStore.delete(id);
PASS event.target.result is undefined
PASS successfullyParsed is true

TEST COMPLETE
